Refactored QC handlers to new multi-file format
